chore: rename stockTransfers transformer file to be camelCase
