+++ packages/app-elements/src/ui/resources/ResourceListItem/transformers/index.tsx
@@ -1,4 +1,4 @@
 export { customerToProps } from './customers'
 export { orderToProps } from './orders'
 export { returnToProps } from './returns'
-export { stockTransferToProps } from './stock_transfers'
+export { stockTransferToProps } from './stockTransfers'
